requirements.txt
ollama==0.3.3
python-dotenv==1.0.0
streamlit
